chore: refactor interaction create handler
