+++ src/modules/events/interactions/processChatInputCommand.ts
@@ -1,0 +1,64 @@
+import { ChatInputCommandInteraction } from "discord.js";
+import { TFunction } from "i18next";
+
+import { BeccaLyria } from "../../../interfaces/BeccaLyria";
+import { currencyListener } from "../../../listeners/currencyListener";
+import { usageListener } from "../../../listeners/usageListener";
+import { beccaErrorHandler } from "../../../utils/beccaErrorHandler";
+import { logActivity } from "../../commands/logActivity";
+import { getSettings } from "../../settings/getSettings";
+
+/**
+ * Handles the logic for slash commands.
+ *
+ * @param {BeccaLyria} Becca Becca's discord instance.
+ * @param {ChatInputCommandInteraction} interaction The interaction payload from Discord.
+ * @param {TFunction} t Translation function.
+ */
+export const processChatInputCommand = async (
+  Becca: BeccaLyria,
+  interaction: ChatInputCommandInteraction,
+  t: TFunction
+) => {
+  try {
+    await logActivity(Becca, interaction.user.id, "command");
+    const target = Becca.commands.find(
+      (el) => el.data.name === interaction.commandName
+    );
+    if (!target) {
+      interaction.reply({
+        content: t<string, string>("events:interaction.bad", {
+          command: interaction.commandName,
+        }),
+      });
+      return;
+    }
+    if (!interaction.guildId || !interaction.guild) {
+      await interaction.reply({
+        content: t<string, string>("events:interaction.noDms"),
+      });
+      return;
+    }
+    const config = await getSettings(
+      Becca,
+      interaction.guildId,
+      interaction.guild.name
+    );
+    if (!config) {
+      await interaction.reply({
+        content: t<string, string>("events:interaction.noSettings"),
+      });
+      return;
+    }
+    await target.run(Becca, interaction, t, config);
+    await usageListener.run(Becca, interaction);
+    await currencyListener.run(Becca, interaction);
+  } catch (err) {
+    await beccaErrorHandler(
+      Becca,
+      "process chat input command",
+      err,
+      interaction.guild?.name
+    );
+  }
+};
